refactor: 기존의 docker controller 코드를 제거하고 controller2를 controller로 변경 (#152)
* refactor: 기존의 docker controller 코드를 제거하고 controller2를 controller로 변경

* fix: 기본 예제의 project 파일의 형식을 동작하도록 변경
