+++ python/mlad/cli/deploy.py
@@ -13,7 +13,7 @@ from mlad.cli.exceptions import (
     ImageNotFoundError, InvalidProjectKindError, InvalidUpdateOptionError
 )
 
-from mlad.core.docker import controller2 as docker_ctlr
+from mlad.core.docker import controller as docker_ctlr
 from mlad.core.default import project as default_project
 from mlad.core.libs import utils as core_utils
